:sparkles: [#110] Notify camunda after doc creation
