+++ src/bptl/work_units/xential/views.py
@@ -1,10 +1,8 @@
-from urllib.parse import urlparse
 from uuid import UUID
 
 from django.http import HttpResponse, HttpResponseRedirect
 from django.views import View
 
-from defusedxml import minidom
 from rest_framework import status, views
 from rest_framework.request import Request
 from rest_framework.response import Response
@@ -15,22 +13,12 @@ from bptl.work_units.xential.serializers import CallbackDataSerializer
 from ...tasks.base import check_variable
 from .client import DRC_ALIAS, XENTIAL_ALIAS, get_client
 from .handlers import on_document_created
-
-
-def parse_xml(raw_xml: str) -> dict:
-    parsed_xml = minidom.parseString(raw_xml)  # minidom from defusedxml
-    extracted_data = {}
-
-    document_node = parsed_xml.getElementsByTagName("document")
-    extracted_data["document"] = document_node[0].firstChild.nodeValue
-
-    ticket_node = parsed_xml.getElementsByTagName("bptlTicketUuid")
-    extracted_data["bptl_ticket_id"] = ticket_node[0].firstChild.nodeValue
-
-    return extracted_data
+from .utils import get_xential_base_url, parse_xml
 
 
 class DocumentCreationCallbackView(views.APIView):
+    permission_classes = []
+
     def post(self, request: Request) -> Response:
         # The callback sends the base64 encoded document and the BPTL ticket ID as XML.
         callback_data = parse_xml(request.data)
@@ -50,7 +38,7 @@ class DocumentCreationCallbackView(views.APIView):
         document_properties["inhoud"] = serializer.validated_data["document"]
 
         drc_client = get_client(task, DRC_ALIAS)
-        drc_client.post("/enkelvoudiginformatieobjecten", json=document_properties)
+        drc_client.post("enkelvoudiginformatieobjecten", json=document_properties)
 
         # Notify camunda that the document has been created
         on_document_created(task)
@@ -83,8 +71,3 @@ class InteractiveDocumentView(View):
 
         # Redirect the user to the Xential URL to interactively create a document
         return HttpResponseRedirect(redirect_to=xential_url)
-
-
-def get_xential_base_url(api_root: str) -> str:
-    parsed_url = urlparse(api_root)
-    return f"{parsed_url.scheme}://{parsed_url.netloc}"
